pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "enhancekit"
version = "0.1.0"
description = "A unified interface for deep learning image restoration and enhancement."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Aymen SEKHRI", email = "asekhri.pro@gmail.com"}]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "Pillow>=9.0.0",
    "pydantic>=2.0.0",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy",
]

[project.scripts]
enhancekit = "enhancekit.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
